Rename lib.go -> lib_libwasmvm.go
